Adjust the Swagger Docs and fix a deprecation in SeedLoader

+++ src/StoryGen/Services/SeedLoader.php
@@ -331,8 +331,9 @@ class SeedLoader
             $combined = [];
 
             // Add some events without modifiers
-            $combined = array_merge($combined, array_slice($withoutModifiers, 0, count($withoutModifiers) * 2 / 3));
-
+            $combined = array_merge($combined, array_slice($withoutModifiers, 0,
+                (int)(count($withoutModifiers) * 2 / 3)
+            ));
             // Add some events with modifiers
             foreach ($verbs as $verb) {
                 foreach ($objects as $object) {
